Only update database when there are actual changes
Containers are now only updated in the database when the container has
changed on the system. As the container name and status are the only
mutable container properties (that we track, e.g. network is mutable but
not tracked), these are the only ones we check for, and also the only
triples we actually update.

This hopefully reduces the load on the database and the amount of deltas
which are sent.